adk-sql-agent/pyproject.toml
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "adk-sql-agent"
version = "0.1.0"
description = "An ADK agent that can run queries on an ephemeral SQLite database"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fsspec[gcs]>=2025.5.1",
    "google-adk>=1.0.0",
    "opentelemetry-exporter-gcp-logging>=1.9.0a0",
    "opentelemetry-exporter-gcp-monitoring>=1.9.0a0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.33.1",
    "opentelemetry-instrumentation-aiohttp-client>=0.54b1",
    "opentelemetry-instrumentation-google-genai>=0.2b0",
    "opentelemetry-instrumentation-httpx>=0.54b1",
    "opentelemetry-instrumentation-requests>=0.54b1",
    "opentelemetry-instrumentation-sqlite3>=0.54b1",
    "opentelemetry-instrumentation-vertexai>=2.0b0",
    "rich>=14.0.0",
]

[dependency-groups]
dev = ["ipython>=9.3.0", "pillow>=11.2.1"]
